+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to the
 ### Changed
 - Interpreter checks whether first command-line arg is a valid file before attempting to read the program from it (if not, it's treated as just another argument and the program is read from stdin)
 - Cycle and related builtins, when given empty list, return empty list instead of erroring
+- Flipped order of arguments for comparison builtins
 
 ## 0.0.3 - 2024-06-01